优化用户缓存过期时间

+++ wind-server/wind-common/wind-common-utils/src/main/java/com/sunseagear/common/utils/CacheUtils.java
@@ -53,7 +53,7 @@ public class CacheUtils {
      * @param timeout  时间
      * @param timeUnit 时间颗粒度
      */
-    public static <T> void setCacheObject(final String key, final T value, final Integer timeout, final TimeUnit timeUnit) {
+    public static <T> void setCacheObject(final String key, final T value, final long timeout, final TimeUnit timeUnit) {
         redisTemplate.opsForValue().set(key, value, timeout, timeUnit);
     }
